cast to double to get actual float delta, rebuild ndlls

+++ project/src/backend/sdl3/SDLApplication.cpp
@@ -136,7 +136,7 @@ namespace lime {
 					currentUpdate = SDL_GetTicksNS ();
 					applicationEvent.type = UPDATE;
 					// convert to ms to keep avoid a breaking change
-					applicationEvent.deltaTime = SDL_NS_TO_MS ((currentUpdate - lastUpdate));
+					applicationEvent.deltaTime = SDL_NS_TO_MS ((double)(currentUpdate - lastUpdate));
 
 					lastUpdate = currentUpdate;
